print_last_digit

+++ 0x02-functions_nested_loops/101-natural.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-int sum, i;
+long unsigned int sum, i;
 for (i = 0; i < 1024; ++i)
 {
 if ((i % 3) == 0 || (i % 5) == 0)
